Fixes + moved classes
