+++ src/main/java/ch/njol/unofficialmonumentamod/features/discordrpc/DiscordRPC.java
@@ -1,11 +1,9 @@
-package ch.njol.unofficialmonumentamod.discordrpc;
+package ch.njol.unofficialmonumentamod.features.discordrpc;
 
 import ch.njol.unofficialmonumentamod.UnofficialMonumentaModClient;
-import ch.njol.unofficialmonumentamod.misc.Locations;
-import ch.njol.unofficialmonumentamod.mixins.PlayerListHudAccessor;
+import ch.njol.unofficialmonumentamod.features.locations.Locations;
 import club.minnced.discord.rpc.*;
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.text.Text;
 import net.minecraft.util.Hand;
 
 import java.util.*;
